Create 05_Git_Node_VSCode_install.md
Note that the current Node version is now 10.x. In the future, just changing the line that says `/setup_10.x` to `11.x` will fix this, but that may be a while. Also, even if 9 was installed this way you can type the same two exact commands as before but change it from 9 to 10 and it will auto detect the older client and automatically remove it and update it to 10.