Update 04_updating_terminal.md
Fix for the wr export. Removed the ' and the message for it. The reason that was there was because before we were using just an alias, but here we export a variable to store the path to the windows root, so we can use it in other places in place of ~. Changing this meant we didn't need the ' anymore.